+++ frontend/src/app/modules/admin/asamblea/cierreAsamblea/cierreAsamblea.component.html
@@ -117,88 +117,5 @@
         </div>
     </div>
 
-    <!-- seccion de parametrizacion  -->
-    <div
-        class="ml-0.5  text-2xl font-extrabold text-secondary flex flex-col flex-auto items-start justify-center mt-5 ">
-        <span class="flex justify-start">Parametrización</span>
-    </div>
-    <form class="grid grid-cols-2 gap-4 bg-card py-4 shadow rounded mt-2" action="" [formGroup]="formulario"
-        (ngSubmit)="enviarFormularioParametrizacion()">
 
-        <div class="p-5 py-3">
-            <mat-form-field class="w-full mr-2 ">
-                <mat-label class="font-extrabold"> Acciones En El Mercado:</mat-label>
-                <mat-icon svgIcon="feather:hash"></mat-icon>
-                <input formControlName="numAccMercado" matInput type="number">
-                <mat-error class="mb-2">Este campo es numérico y obligatorio</mat-error>
-            </mat-form-field>
-            <mat-form-field class="w-full mr-2 ">
-                <mat-label class="font-extrabold"> Acciones Utilidades:</mat-label>
-                <mat-icon svgIcon="feather:hash"></mat-icon>
-                <input formControlName="numAccUtilidades" matInput type="number">
-                <mat-error class="mb-2">Este campo es numérico y obligatorio </mat-error>
-            </mat-form-field>
-            <mat-form-field class="w-full mr-2 ">
-                <mat-label class="font-extrabold"> Participación Por Acción:</mat-label>
-                <mat-icon svgIcon="feather:hash"></mat-icon>
-                <input formControlName="participacionAccion" matInput type="number">
-                <mat-error class="mb-2">Este campo es numérico y obligatorio </mat-error>
-            </mat-form-field>
-            <mat-form-field class="w-full mr-2">
-                <mat-label class=" font-extrabold mt-2">Tipo de
-                    Asamblea</mat-label>
-                <mat-select formControlName="pagoUtilidad">
-                    <mat-option [value]=1>Un Pago</mat-option>
-                    <mat-option [value]=2>Dos Pagos</mat-option>
-                    <mat-option [value]=3>Tres Pagos</mat-option>
-                </mat-select>
-                <mat-error class="mb-2">Este campo es obligatorio</mat-error>
-            </mat-form-field>
-
-            <div formArrayName="numPagos" class="flex flex-row space-x-3 justify-center items-center">
-                <div *ngFor="let control of numPagos.controls; let i = index">
-                    <mat-form-field>
-                        <input matInput [formControlName]="i" placeholder="pago {{ i + 1 }}" type="number">
-                        <mat-icon svgIcon="feather:percent"></mat-icon>
-                    </mat-form-field>
-                </div>
-            </div>
-            <div class="flex flex-row justify-center">
-
-                <mat-error class="mb-2"  *ngIf="numPagos.errors?.sumExcedida">Maximo permitido de 100%</mat-error>
-                <!-- <mat-error class="mb-2"  *ngIf="camposDinamicos.errors?.sumNoIgualA100">Debe ser del 100%</mat-error> -->
-            </div>
-
-        </div>
-        <!-- seccion dos -->
-        <div class="p-5 py-3">
-            <mat-form-field class="w-full mr-2 ">
-                <mat-label class="font-extrabold"> Valor Nominal Por Acción:</mat-label>
-                <mat-icon svgIcon="feather:hash"></mat-icon>
-                <input formControlName="valNomAccion" matInput type="number">
-                <mat-error class="mb-2">Este campo es numérico y obligatorio </mat-error>
-            </mat-form-field>
-            <mat-form-field class="w-full mr-2 ">
-                <mat-label class="font-extrabold"> Valor Intrínseco:</mat-label>
-                <mat-icon svgIcon="feather:hash"></mat-icon>
-                <input formControlName="valIntrinseco" matInput type="number">
-                <mat-error class="mb-2">Este campo es numérico y obligatorio </mat-error>
-            </mat-form-field>
-            <mat-form-field class="w-full mr-2 ">
-                <mat-label class="font-extrabold"> Dividendos Por Participación:</mat-label>
-                <mat-icon svgIcon="feather:hash"></mat-icon>
-                <input formControlName="divParticipacion" matInput type="number">
-                <mat-error class="mb-2">Este campo es numérico y obligatorio </mat-error>
-            </mat-form-field>
-            <div class="w-full justify-start items-end">
-
-                <button mat-flat-button color="accent" class="w-full sm:mt-7 space-x-1" type="submit"
-                    [disabled]="formulario.invalid">
-                    <mat-icon svgIcon="feather:save"></mat-icon>
-                    <span>Guardar</span></button>
-            </div>
-
-        </div>
-
-    </form>
 </div>
