+++ frontend/src/app/modules/admin/financiero/financiero.component.html
@@ -1,0 +1,148 @@
+<fuse-loading-bar></fuse-loading-bar>
+
+<div class="flex-auto p-6 sm:p-10">
+    <!-- titulos -->
+    <div class="text-4xl font-extrabold flex flex-col flex-auto items-center justify-center ">Modulo Financiero</div>
+    <div class="ml-0.5 mb-5 text-4xl font-extrabold text-secondary flex flex-col flex-auto items-center justify-center">
+        Dividendos</div>
+
+
+
+    <!-- alertas -->
+    <div class="flex justify-end fixed top-7 right-9 mt-13 w-full z-50 animated-box">
+
+        <div class="w-100 mb-4 ">
+
+            <div class="fixed  top-5 right-9 mt-5 w-80 z-50 animated-box" *ngIf="showSuccesAlert">
+                <fuse-alert [type]="'success'" [appearance]="'outline'" class="animated-box-content" @zoomIn>
+                    <span fuseAlertTitle>Información</span>
+                    ¡Transacción exitosa!
+                </fuse-alert>
+            </div>
+
+            <div class="fixed  top-5 right-9 mt-5 w-80 z-50 animated-box" *ngIf="showFailedAlert">
+                <fuse-alert [type]="'warn'" [appearance]="'outline'" class="animated-box-content" @shake>
+                    <span fuseAlertTitle>Información</span>
+                    ¡Transacción no exitosa!
+                </fuse-alert>
+            </div>
+        </div>
+    </div>
+    <!-- fin alertas -->
+
+    <!-- Tabla con registros -->
+
+    <!-- seccion de parametrizacion  -->
+    <div
+        class="ml-0.5  text-2xl font-extrabold text-secondary flex flex-col flex-auto items-start justify-center mt-5 ">
+        <span class="flex justify-start">Parametrización</span>
+    </div>
+    <form class="grid grid-cols-2 gap-4 bg-card py-4 shadow rounded mt-2" action="" [formGroup]="formulario"
+        (ngSubmit)="enviarFormularioParametrizacion()">
+
+        <div class="p-5 py-3">
+            <mat-form-field class="w-full mr-2 ">
+                <mat-label class="font-extrabold"> Acciones En El Mercado:</mat-label>
+                <mat-icon svgIcon="feather:hash"></mat-icon>
+                <input formControlName="numAccMercado" matInput type="text">
+                <mat-error class="mb-2">Este campo es numérico y obligatorio</mat-error>
+            </mat-form-field>
+            <mat-form-field class="w-full mr-2 ">
+                <mat-label class="font-extrabold"> Acciones Utilidades:</mat-label>
+                <mat-icon svgIcon="feather:hash"></mat-icon>
+                <input formControlName="numAccUtilidades" matInput type="text">
+                <mat-error class="mb-2">Este campo es numérico y obligatorio </mat-error>
+            </mat-form-field>
+            <mat-form-field class="w-full mr-2 ">
+                <mat-label class="font-extrabold"> Participación Por Acción:</mat-label>
+                <mat-icon svgIcon="feather:hash"></mat-icon>
+                <input formControlName="participacionAccion" matInput type="text">
+                <mat-error class="mb-2">Este campo es numérico y obligatorio </mat-error>
+            </mat-form-field>
+            <mat-form-field class="w-full mr-2">
+                <mat-label class=" font-extrabold mt-2">Tipo de
+                    Asamblea</mat-label>
+                <mat-select formControlName="pagoUtilidad">
+                    <mat-option [value]=1>Un Pago</mat-option>
+                    <mat-option [value]=2>Dos Pagos</mat-option>
+                    <mat-option [value]=3>Tres Pagos</mat-option>
+                </mat-select>
+                <mat-error class="mb-2">Este campo es obligatorio</mat-error>
+            </mat-form-field>
+
+            <div formArrayName="numPagos" class="flex flex-row space-x-3 justify-center items-center">
+                <div *ngFor="let control of numPagos.controls; let i = index">
+                    <mat-form-field>
+                        <input matInput [formControlName]="i" placeholder="pago {{ i + 1 }}" type="number">
+                        <mat-icon svgIcon="feather:percent"></mat-icon>
+                    </mat-form-field>
+                </div>
+            </div>
+            <div class="flex flex-row justify-center">
+
+                <mat-error class="mb-2"  *ngIf="numPagos.errors?.sumExcedida">Maximo permitido de 100%</mat-error>
+                <!-- <mat-error class="mb-2"  *ngIf="camposDinamicos.errors?.sumNoIgualA100">Debe ser del 100%</mat-error> -->
+            </div>
+
+        </div>
+        <!-- seccion dos -->
+        <div class="p-5 py-3">
+            <mat-form-field class="w-full mr-2 ">
+                <mat-label class="font-extrabold"> Valor Nominal Por Acción:</mat-label>
+                <mat-icon svgIcon="feather:hash"></mat-icon>
+                <input formControlName="valNomAccion" matInput type="text">
+                <mat-error class="mb-2">Este campo es numérico y obligatorio </mat-error>
+            </mat-form-field>
+            <mat-form-field class="w-full mr-2 ">
+                <mat-label class="font-extrabold"> Valor Intrínseco:</mat-label>
+                <mat-icon svgIcon="feather:hash"></mat-icon>
+                <input formControlName="valIntrinseco" matInput type="text">
+                <mat-error class="mb-2">Este campo es numérico y obligatorio </mat-error>
+            </mat-form-field>
+            <mat-form-field class="w-full mr-2 ">
+                <mat-label class="font-extrabold"> Dividendos Por Participación:</mat-label>
+                <mat-icon svgIcon="feather:hash"></mat-icon>
+                <input formControlName="divParticipacion" matInput type="text">
+                <mat-error class="mb-2">Este campo es numérico y obligatorio </mat-error>
+            </mat-form-field>
+            <div class="w-full justify-start items-end">
+
+                <button mat-flat-button color="accent" class="w-full sm:mt-7 space-x-1" type="submit"
+                    [disabled]="formulario.invalid">
+                    <mat-icon svgIcon="feather:save"></mat-icon>
+                    <span>Guardar</span></button>
+            </div>
+
+        </div>
+
+    </form>
+
+
+            <!-- modulo dividendos -->
+            <p class="my-2 text-xl text-gray-500 font-semibold mt-5">Reportes Dividendos</p>
+            <div class="flex-auto pt-5 bg-card shadow rounded overflow-hidden  mb-3 py-4">
+                <div class="flex ">
+
+                    <div class="flex flex-col w-full justify-center space-y-4 py-2 px-10 items-center">
+                        <div class="w-full flex flex-row justify-between">
+
+                            <div class="w-100 mr-5">
+                                <mat-form-field class="w-full mr-2">
+                                    <mat-label class=" font-extrabold mt-2">Año del reporte</mat-label>
+                                    <mat-select [formControl]="anioSeleccionado" >
+                                        <mat-option *ngFor="let anio of pruebasAnios" [value]="anio">{{anio}}</mat-option>
+                                    </mat-select>
+                                </mat-form-field>
+                            </div>
+                            <div class="mt-3">
+                                <button mat-fab [color]="'accent'" [disabled]="this.anioSeleccionado.invalid"
+                                    (click)="obtenerReporteDivididendo()"><mat-icon
+                                        svgIcon="heroicons_solid:document-arrow-down"></mat-icon></button>
+
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+
+</div>
