+++ frontend/src/app/modules/admin/reportes/reportes/reportes.component.html
@@ -278,32 +278,7 @@
                 </div>
             </div>
         </div>
-        <!-- modulo dividendos -->
-        <p class="my-2 text-xl text-gray-500 font-semibold mt-5">Reportes Dividendos</p>
-        <div class="flex-auto pt-5 bg-card shadow rounded overflow-hidden  mb-3 py-4">
-            <div class="flex ">
 
-                <div class="flex flex-col w-full justify-center space-y-4 py-2 px-10 items-center">
-                    <div class="w-full flex flex-row justify-between">
-
-                        <div class="w-100 mr-5">
-                            <mat-form-field class="w-full mr-2">
-                                <mat-label class=" font-extrabold mt-2">Año del reporte</mat-label>
-                                <mat-select [formControl]="anioSeleccionado" >
-                                    <mat-option *ngFor="let anio of pruebasAnios" [value]="anio">{{anio}}</mat-option>
-                                </mat-select>
-                            </mat-form-field>
-                        </div>
-                        <div class="mt-3">
-                            <button mat-fab [color]="'accent'" [disabled]="this.anioSeleccionado.invalid"
-                                (click)="obtenerReporteDivididendo()"><mat-icon
-                                    svgIcon="heroicons_solid:document-arrow-down"></mat-icon></button>
-
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
 
         <p class="my-2 text-xl text-gray-500 font-semibold mt-5">Módulo Móvil</p>
         <div class="flex-auto pt-5 bg-card shadow rounded overflow-hidden  mb-3 py-4">
